+++ src/servers/box2d_project_settings.cpp
@@ -8,6 +8,7 @@ constexpr char RUN_ON_SEPARATE_THREAD[] = "physics/2d/run_on_separate_thread";
 constexpr char MAX_THREADS[] = "threading/worker_pool/max_threads";
 constexpr char POSITION_ITERATIONS[] = "physics/box_2d/solver/position_iterations";
 constexpr char VELOCITY_ITERATIONS[] = "physics/box_2d/solver/velocity_iterations";
+constexpr char SCALING_FACTOR[] = "physics/box_2d/scaling_factor";
 
 void register_setting(
 		const String &p_name,
@@ -66,6 +67,7 @@ void register_setting_ranged(
 void Box2DProjectSettings::register_settings() {
 	register_setting_ranged(VELOCITY_ITERATIONS, 8, U"2,16,or_greater");
 	register_setting_ranged(POSITION_ITERATIONS, 3, U"1,16,or_greater");
+	register_setting_ranged(SCALING_FACTOR, 100.0f, U"1,100,or_greater");
 }
 
 template <typename TType>
@@ -94,4 +96,8 @@ int Box2DProjectSettings::get_position_iterations() {
 
 int Box2DProjectSettings::get_velocity_iterations() {
 	return get_setting<int>(VELOCITY_ITERATIONS);
+}
+
+float Box2DProjectSettings::get_scaling_factor() {
+	return get_setting<float>(SCALING_FACTOR);
 }